{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "{{cookiecutter.version}}"
description = ""
authors = ["Kumar Pratik <kumar.pratik@senecaglobal.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.39"
requests = "^2.28.1"
python-jose = "^3.3.0"
aiohttp = "^3.8.1"
PyMySQL = "^1.0.2"
python-multipart = "^0.0.5"
redis = "^4.5.1"
more-itertools = "^9.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
black = "^22.10.0"
pylint = "^2.15.3"
alembic = "^1.8.1"
flake8 = "^6.0.0"
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
